backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
pypdf==4.3.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
tqdm==4.66.4
google-generativeai==0.7.2
orjson==3.10.6